[front] refactor(mcp): migrate file_generation server to new structure
- Move file_generation server from legacy location to lib/api/actions/servers/
- Split into metadata.ts, helpers.ts, tools/index.ts, and index.ts
- Use createToolsRecord and buildTools pattern for type-safe tool definitions
- Update constants.ts to use metadata pattern
- Update servers/index.ts to import from new location
