+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -16,7 +16,6 @@ import { default as dataSourcesFileSystemServer } from "@app/lib/actions/mcp_int
 import { default as dataWarehousesServer } from "@app/lib/actions/mcp_internal_actions/servers/data_warehouses";
 import { default as databricksServer } from "@app/lib/actions/mcp_internal_actions/servers/databricks";
 import { default as deepDiveServer } from "@app/lib/actions/mcp_internal_actions/servers/deep_dive";
-import { default as generateFileServer } from "@app/lib/actions/mcp_internal_actions/servers/file_generation";
 import { default as freshserviceServer } from "@app/lib/actions/mcp_internal_actions/servers/freshservice";
 import { default as frontServer } from "@app/lib/actions/mcp_internal_actions/servers/front";
 import { default as gmailServer } from "@app/lib/actions/mcp_internal_actions/servers/gmail";
@@ -69,6 +68,7 @@ import {
 } from "@app/lib/actions/types/guards";
 import { default as agentCopilotAgentStateServer } from "@app/lib/api/actions/servers/agent_copilot_agent_state";
 import { default as agentCopilotContextServer } from "@app/lib/api/actions/servers/agent_copilot_context";
+import { default as fileGenerationServer } from "@app/lib/api/actions/servers/file_generation";
 import { default as githubServer } from "@app/lib/api/actions/servers/github";
 import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
 import { default as imageGenerationServer } from "@app/lib/api/actions/servers/image_generation";
@@ -124,7 +124,7 @@ export async function getInternalMCPServer(
     case "sound_studio":
       return soundStudio(auth, agentLoopContext);
     case "file_generation":
-      return generateFileServer(auth, agentLoopContext);
+      return fileGenerationServer(auth, agentLoopContext);
     case "interactive_content":
       return interactiveContentServer(auth, agentLoopContext);
     case "query_tables_v2":
